Added pydantic and moved models into folder
