+++ DiscTrackerDjango/disctracker/items/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from .models import Item, PriceHistory
+from items.models.db_models import Item, PriceHistory
 
 # Register your models here.
 admin.site.register(Item)
