+++ DiscTrackerDjango/disctracker/items/models/pydantic_models.py
@@ -1,0 +1,14 @@
+from pydantic import BaseModel
+
+class ItemDetail(BaseModel):
+    boxId: str
+    boxName: str
+    sellPrice: float 
+    exchangePrice: float
+    cashPrice: float
+
+class CexItemData(BaseModel):
+    data: ItemDetail
+
+class CexItemResponse(BaseModel):
+    response: CexItemData
