+++ DiscTrackerDjango/disctracker/items/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from . import views
+from items import views
 
 app_name="items"
 urlpatterns = [
