+++ requirements.txt
@@ -1,8 +1,11 @@
+annotated-types==0.7.0
+asgiref==3.8.1
 beautifulsoup4==4.12.3
 blinker==1.9.0
 certifi==2024.8.30
 charset-normalizer==3.3.2
 click==8.1.8
+Django==5.1.5
 Flask==3.1.0
 idna==3.8
 iniconfig==2.0.0
@@ -14,6 +17,8 @@ packaging==24.1
 pandas==2.2.2
 pluggy==1.5.0
 psycopg2-binary==2.9.10
+pydantic==2.10.6
+pydantic_core==2.27.2
 pytest==8.3.2
 python-dateutil==2.9.0.post0
 python-dotenv==1.0.1
@@ -23,6 +28,7 @@ schedule==1.2.2
 six==1.16.0
 soupsieve==2.6
 SQLAlchemy==2.0.37
+sqlparse==0.5.3
 typing_extensions==4.12.2
 tzdata==2024.1
 urllib3==2.2.2
